+++ crates/dmm-lib/src/lib.rs
@@ -3,6 +3,7 @@ pub(crate) mod ch9325;
 pub(crate) mod ch9329;
 pub mod cp2110;
 pub mod error;
+pub mod export;
 pub mod flags;
 pub mod measurement;
 pub mod mock;